Reactor upload function
This will refactor upload function.
